Version 0.2.2
Split Match Type and Stipulation to two dropdowns so you can have Tag Team Hell in a Cells etc
Added more Stipulations
Added Royal Rumble Match Type. No participants are needed to be selected when choosing this type. Just choose the winner and it will reflect in Histories
Added Money in the Bank Stipulation. To use create a Money in the Bank Briefcase as a title and put it on the line.
Added Trios, 5-Ways and 6-Way matches that allow up to 6 participant selections.
Added new information to the Promotion Snapshot screen
Now when you select a title in Titles Tab it will show the current champion